requirements.txt
google-genai
python-dotenv
google-adk
